+++ changelog.md
@@ -6,6 +6,7 @@ Unreleased features and fixes can be viewed on GitHub. To do this, click on [Unr
 
 ### Added
 - Multi-language icon URL setting for payment selection description [(#80)](https://github.com/RobinTheHood/modified-stripe/pull/80)
+- Add option to reset auto-increment order number after deleting temporary orders ([#84](https://github.com/RobinTheHood/modified-stripe/issues/84))
 
 ## [0.12.0] - 2025-05-19
 ### Added